pyproject.toml
[project]
name = "insight-blueprint"
version = "0.4.4"
description = "MCP server for data science analysis design management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "insight-blueprint contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Framework :: FastAPI",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.10",
    "ruamel.yaml>=0.18",
    "click>=8.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.41.0",
    "packaging>=23.0",
    "filelock>=3.24.2",
]

[project.scripts]
insight-blueprint = "insight_blueprint.cli:main"

[project.urls]
Homepage = "https://github.com/etoyama/insight-blueprint"
Repository = "https://github.com/etoyama/insight-blueprint"
"Bug Tracker" = "https://github.com/etoyama/insight-blueprint/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "ty>=0.0.1",
    "poethepoet>=0.31",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/insight_blueprint"]
artifacts = ["src/insight_blueprint/static/**"]

[tool.hatch.build.targets.sdist]
artifacts = ["src/insight_blueprint/static/**"]

# ============================================================
# ruff
# ============================================================
[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [".claude/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

# ============================================================
# pytest
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests", "."]
addopts = "-v --tb=short"

# ============================================================
# poethepoet
# ============================================================
[tool.poe.tasks.lint]
shell = "uv run ruff check . && uv run ruff format --check ."

[tool.poe.tasks.format]
shell = "uv run ruff check --fix . && uv run ruff format ."

[tool.poe.tasks.typecheck]
cmd = "uv run ty check src/"

[tool.poe.tasks.test]
cmd = "uv run pytest -v"

[tool.poe.tasks.install-frontend]
cmd = "npm --prefix frontend ci"
help = "Install frontend dependencies"

[tool.poe.tasks.build-frontend]
sequence = ["install-frontend", {cmd = "npm --prefix frontend run build"}]
help = "Install deps and build frontend assets"

[tool.poe.tasks.build]
sequence = ["build-frontend"]

[tool.poe.tasks.all]
sequence = ["lint", "typecheck", "test"]

[tool.poe.tasks.ci]
sequence = ["lint", "typecheck", "test", "build-frontend"]
help = "Run the full CI pipeline locally (Python + frontend)"

[tool.poe.tasks.verify-wheel]
cmd = "uv run python scripts/verify_wheel.py"
help = "Verify wheel contains frontend assets and py.typed"

[tool.poe.tasks.check-tag-version]
cmd = "uv run python scripts/check_tag_version.py"
help = "Check git tag matches pyproject.toml version"

[tool.poe.tasks.release-dry-run]
sequence = ["build-frontend", {shell = "rm -rf dist/ && uv build"}, "verify-wheel", {shell = "uvx --from twine twine check dist/*"}]
help = "Full release dry-run: build frontend, build wheel, verify, twine check"

[dependency-groups]
dev = [
    "httpx>=0.27",
    "marimo>=0.21.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
    "ty>=0.0.17",
]
